requirements/testing.txt
-r base.txt

mock
aloe
aloe_django
aloe_webdriver

black

vcrpy==1.11.0

pytest
pytest-django
pytest-pep8
pytest-flakes
pytest-cov
